requirement.txt
# Climate Comrade Dependencies
openai>=1.0.0
pyserial>=3.5
python-dotenv>=1.0.0

# Development dependencies
black>=23.0.0
pylint>=2.17.0
mypy>=1.0.0
pytest>=7.0.0